# Using `Array.Copy` to make array copy faster in StylusPointCollection